State: Cloning code repos.
The Ansible playbook now goes as far as cloning code repos.

Renamed 'setup.sh' to 'ram'. Plan to add it to PATH so members have a
convenient command for managing their virtual machine and projects.

Moved Ansible roles around. Combined embedded and software roles,
because we are aiming to reduce the gap between the software and
embedded teams, so this organization makes more sense. Moved the SSH
key generation to a role called 'postinstall'. Postinstall should be
run on the end-user's machine when they setup their environment. It
should not be run when creating a base VM image to distribute to users.