Reword SSH pubkey uploading instructions
We print two messages. The first includes the GitLab key settings URL
and the SSH public key. The second message prints instructions and
pauses, waiting for the user to press enter/return. This commit
adds text to the second message to mention that the instructions are
printed above, as it is likely the user is looking at the most
recently printed message.