Refactor
Base image commands now under `ram image` subcommand for the sake of
organization and warning end-users away from them.

Now 3 Ansible roles, split by when they should be run.

"software" was renamed to "core".

The new "base" role (that is not run by 'ram setup') allows for things
that should be included in the base image but can be removed freely by end-users.